// src/app/shared/models/section10-model.ts
import { Section0Model } from "./section0-model";

export class Section10Model extends Section0Model {
  induccionCorporativa: string = "";
  grupoFamiliar: string = "";
  parentescoPrincipal: string = "";
  observaciones: string = "";
  consOtraInfo?: string;
  consRedSocial?: number;
  userRedSocial?: string = "";
  parentesco?: number;
  tipoIdentificacionPrincipal?: string = "";
}