// src/app/shared/models/section7-model.ts
import { Section0Model } from "./section0-model";

export class Section7Model extends Section0Model {
  consInfoAdicional?: string = "";
  ninosACargo: number;
  adultosACargo: number;
  nivelAcademico: number;
  fechaGrado: number;  
  institucion: number;
  titulo: number;
  actualmente: number;
}