// src/app/shared/models/section9-model.ts
import { Section0Model } from "./section0-model";

export class Section9Model extends Section0Model {
  orgInternPregunta1: string = "";
  cargoPoliPregunta2: string = "";
  descCargoPublico: string = "";
  fechaInicio: 1538424211000;
  fechaFin: 1540238620000;
  actualmente: number = 0;
  idProspecto: number;
}